Arm/alternative: hide casting away of const
We want to patch the original call site, which memory-wise is unrelated
to the struct alt_instr * that we start from.

No functional change intended.

+++ xen/arch/arm/alternative.c
@@ -134,7 +134,7 @@ static int __apply_alternatives(const struct alt_region *region,
             BUG_ON(alt->repl_len != alt->orig_len);
 
         origptr = ALT_ORIG_PTR(alt);
-        updptr = (void *)origptr + update_offset;
+        updptr = (void *)(unsigned long)origptr + update_offset;
 
         nr_inst = alt->orig_len / ARCH_PATCH_INSN_SIZE;
